Do not release pre-releases by accident
