+++ jobs/microk8s/release-to-beta.py
@@ -3,6 +3,7 @@
 import os
 from snapstore import Microk8sSnap
 from configbag import get_tracks
+from utils import upstream_release
 
 
 # Set this to 'no' if you are sure you want to release
@@ -47,6 +48,10 @@ if __name__ == '__main__':
     print("Dry run is set to '{}'.".format(dry_run))
     for track in tracks_requested:
         print("Looking at track {}".format(track))
+        upstream = upstream_release(track)
+        if not upstream:
+            print("No stable upstream release yet.")
+            continue
         edge_snap = Microk8sSnap(track, 'edge', juju_unit, juju_controller)
         if not edge_snap.released:
             print("Nothing released on {} edge.".format(track))
