+++ jobs/microk8s/update-gh-branches-and-lp-builders.py
@@ -7,24 +7,11 @@ from launchpadlib.launchpad import Launchpad
 from lazr.restfulclient.errors import HTTPError
 from configbag import get_tracks
 from subprocess import check_call, CalledProcessError
+from utils import upstream_release
 
 
 gh_user = os.environ.get('GH_USER')
 gh_token = os.environ.get('GH_TOKEN')
-
-
-def upstream_release(release):
-    """Return the latest stable k8s in the release series"""
-    if release == "latest":
-        release_url = "https://dl.k8s.io/release/stable.txt"
-    else:
-        release_url = "https://dl.k8s.io/release/stable-{}.txt".format(release)
-
-    r = requests.get(release_url)
-    if r.status_code == 200:
-        return r.content.decode().strip()
-    else:
-        None
 
 
 def is_latest(release):
